requirements.txt
streamlit==1.49.0
fastapi==0.116.1
pydantic==2.11.7
uvicorn==0.35.0
mysql-connector-python==9.4.0
pandas==2.3.1
requests==2.32.4
pytest==8.4.2